[bug fix] update download url, requires jq library

+++ mcupdate.sh
@@ -9,6 +9,13 @@
 # alias mcupdate='sudo bash /path/to/mcupdate.sh' (add to bash aliases on ubuntu server)
 ####################################
 
+# CHECK REQUIREMENTS
+if ! command -v jq &> /dev/null
+  then
+    echo "Error: jq library not installed. Please install and try again."
+    exit 1
+fi
+
 # STOP SERVER
 echo stopping server ...
 sudo systemctl stop mcbedrock
@@ -23,7 +30,7 @@ ls -al /home/mcserver/
 
 # DOWNLOAD new version
 echo downloading new version ...
-DOWNLOAD_URL=$(curl -H "Accept-Encoding: identity" -H "Accept-Language: en" -s -L -A "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1; BEDROCK-UPDATER)" https://minecraft.net/en-us/download/server/bedrock/ |  grep -o 'https.*/bin-linux/.*.zip')
+DOWNLOAD_URL="$(curl -s https://net-secondary.web.minecraft-services.net/api/v1.0/download/links | jq -r '.result.links[] | select(.downloadType=="serverBedrockLinux") | .downloadUrl')"
 
 sudo wget -U "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1; BEDROCK-UPDATER)" $DOWNLOAD_URL -O /home/mcserver/minecraft_bedrock/bedrock-server.zip
